Paging

+++ app/src/main/java/com/example/ivor_hu/meizhi/ui/fragment/BaseStuffFragment.java
@@ -29,7 +29,7 @@ import java.util.Date;
 /**
  * Created by ivor on 16-6-21.
  */
-public abstract class BaseStuffFragment extends BaseFragment {
+public abstract class BaseStuffFragment<T, VH> extends BaseFragment {
     protected StuffViewModel mStuffViewModel;
     protected StuffFragmentBinding mBinding;
 
@@ -54,11 +54,6 @@ public abstract class BaseStuffFragment extends BaseFragment {
     @Override
     protected SwipeRefreshLayout getRefreshLayout() {
         return mBinding.stuffRefreshLayout;
-    }
-
-    @Override
-    protected int getLastVisiblePos() {
-        return ((LinearLayoutManager) mLayoutManager).findLastVisibleItemPosition();
     }
 
     @Override
@@ -116,19 +111,23 @@ public abstract class BaseStuffFragment extends BaseFragment {
                     return true;
                 case R.id.context_menu_favor:
                     stuff.setLastChanged(new Date());
-                    mStuffViewModel.insertCollection(stuff);
-                    CommonUtil.makeSnackBar(
-                            getView(),
-                            getContext().getString(R.string.favored),
-                            Snackbar.LENGTH_SHORT);
+                    if (mStuffViewModel != null) {
+                        mStuffViewModel.insertCollection(stuff);
+                        CommonUtil.makeSnackBar(
+                                getView(),
+                                getContext().getString(R.string.favored),
+                                Snackbar.LENGTH_SHORT);
+                    }
                     mode.finish();
                     return true;
                 case R.id.context_menu_unfavor:
-                    mStuffViewModel.deleteCollection(stuff);
-                    CommonUtil.makeSnackBar(
-                            getView(),
-                            getContext().getString(R.string.unfavored),
-                            Snackbar.LENGTH_SHORT);
+                    if (mStuffViewModel != null) {
+                        mStuffViewModel.deleteCollection(stuff);
+                        CommonUtil.makeSnackBar(
+                                getView(),
+                                getContext().getString(R.string.unfavored),
+                                Snackbar.LENGTH_SHORT);
+                    }
                     mode.finish();
                     return true;
                 default:
